started MVP on PlaceOrderActivity
