+++ Android/app/src/main/java/controller/customersystem/PlaceOrderViewInterface.java
@@ -6,4 +6,7 @@ public interface PlaceOrderViewInterface {
     void displayDishesOrdered(HashMap<String, Integer> dishesOrdered);
     void orderSuccessfullyPlaced();
     void setErrorMessage(String message);
+
+    void setDishNamePickerMaxValue(int size);
+    void setDisplayedDishNames(String[] dishNames);
 }
