--- Android/app/src/main/java/use_case/boundary/output/MenuOutputBoundary.java
@@ -1,9 +1,0 @@
-package use_case.boundary.output;
-
-public interface MenuOutputBoundary {
-    void setDishNamePickerMaxValue(int size);
-
-    void setDisplayedDishNames(String[] dishNames);
-
-    void updateDishesOrdered(String dishName, int dishQuantity);
-}
