+++ Android/app/src/main/java/use_case/customersystem/PlaceOrderOutputBoundary.java
@@ -1,5 +1,9 @@
 package use_case.customersystem;
 
 public interface PlaceOrderOutputBoundary {
+    void setDishNamePickerMaxValue(int size);
 
+    void setDisplayedDishNames(String[] dishNames);
+
+    void updateDishesOrdered(String dishName, int dishQuantity);
 }
